project/backend/requirements.txt
Flask==2.3.3
Flask-CORS==4.0.0
requests==2.31.0
python-dotenv==1.0.0
SpeechRecognition==3.10.0
google-generativeai==0.3.2
Pillow==10.0.1
pyttsx3==2.90
googletrans==4.0.0rc1
keyboard==0.13.5